refactor(internal): moved BackgroundTaskService to the internal package to allow access to first-party plugins
