+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/DataCollectorTest.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android
 import android.content.Context
 import android.content.res.Configuration
 import android.content.res.Resources
+import com.bugsnag.android.internal.BackgroundTaskService
 import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
