+++ bugsnag-android-core/src/main/java/com/bugsnag/android/InternalReportDelegate.java
@@ -3,8 +3,10 @@ package com.bugsnag.android;
 import static com.bugsnag.android.DeliveryHeadersKt.HEADER_INTERNAL_ERROR;
 import static com.bugsnag.android.SeverityReason.REASON_UNHANDLED_EXCEPTION;
 
+import com.bugsnag.android.internal.BackgroundTaskService;
 import com.bugsnag.android.internal.ImmutableConfig;
 import com.bugsnag.android.internal.JsonHelper;
+import com.bugsnag.android.internal.TaskType;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
