+++ bugsnag-android-core/src/main/java/com/bugsnag/android/TrackerModule.kt
@@ -1,5 +1,6 @@
 package com.bugsnag.android
 
+import com.bugsnag.android.internal.BackgroundTaskService
 import com.bugsnag.android.internal.dag.ConfigModule
 import com.bugsnag.android.internal.dag.DependencyModule
