+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ConfigChangeReceiverTest.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android
 import android.content.Context
 import android.content.res.Configuration
 import android.content.res.Resources
+import com.bugsnag.android.internal.BackgroundTaskService
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
